19-join시-login-기능이-함께-동작하는-문제-js
1. join-login 동시 실행 문제 해결
2. join/login/logout 버튼과 팝업창 내 버튼 실행 함수 분리
   - selectFunction
   - buttonFunction
