--- pctc-front/src/component/signin.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-  <meta charset="UTF-8">
-  <title>Sign in</title>
-</head>
-<body>
-  
-</body>
-</html>
